+++ src/main/java/com/github/cc3002/finalreality/model/character/AbstractCharacter.java
@@ -1,9 +1,12 @@
 package com.github.cc3002.finalreality.model.character;
 
 
+import java.util.Objects;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.TimeUnit;
+
 import org.jetbrains.annotations.NotNull;
 
 /**
@@ -15,10 +18,11 @@ import org.jetbrains.annotations.NotNull;
 public abstract class AbstractCharacter implements ICharacter {
 
   protected final BlockingQueue<ICharacter> turnsQueue;
-  public final String name;
+  protected final String name;
   public int points;
   public final int defense;
   protected ScheduledExecutorService scheduledExecutor;
+  public boolean alive;
 
   protected AbstractCharacter(@NotNull BlockingQueue<ICharacter> turnsQueue,
                               @NotNull String name, int points, int defense) {
@@ -26,13 +30,9 @@ public abstract class AbstractCharacter implements ICharacter {
     this.name = name;
     this.points = points;
     this.defense = defense;
+    this.alive = true;
   }
 
-  @Override
-  public void waitTurn() {
-    scheduledExecutor = Executors.newSingleThreadScheduledExecutor();
-
-  }
   /**
    * Adds this character to the turns queue.
    */
@@ -40,6 +40,11 @@ public abstract class AbstractCharacter implements ICharacter {
   public void addToQueue() {
     turnsQueue.add(this);
     scheduledExecutor.shutdown();
+  }
+
+  @Override
+  public void waitTurn() {
+    scheduledExecutor = Executors.newSingleThreadScheduledExecutor();
   }
 
   @Override
@@ -52,9 +57,45 @@ public abstract class AbstractCharacter implements ICharacter {
     return points;
   }
 
+  public void setPoints(int points){
+    this.points = points;
+  }
+
   @Override
   public int getDefense() {
     return defense;
   }
-}
 
+  @Override
+  public int hashCode() {
+    return Objects.hash(AbstractCharacter.class, getName());
+  }
+
+  @Override
+  public boolean equals(final Object o) {
+    if (this == o) {
+      return true;
+    }
+    if (!(o instanceof AbstractCharacter)) {
+      return false;
+    }
+    final AbstractCharacter that = (AbstractCharacter) o;
+    return getName().equals(that.getName());
+  }
+
+
+
+
+
+
+  public void attackedBy(int damage){
+    if (damage >= this.getPoints()){
+      this.alive = false;
+    }
+    this.setPoints(this.getPoints()-damage);
+  }
+
+  public boolean isAlive(){
+    return this.alive;
+  }
+}