+++ src/main/java/com/github/cc3002/finalreality/model/character/player/IPlayerCharacter.java
@@ -1,6 +1,7 @@
 package com.github.cc3002.finalreality.model.character.player;
 
 import com.github.cc3002.finalreality.model.weapon.AbstractWeapon;
+import com.github.cc3002.finalreality.model.weapon.IWeapon;
 
 /**
  * This represents a player character from the game.
@@ -15,10 +16,11 @@ public interface IPlayerCharacter {
     /**
      * Equips the character with a weapon.
      */
-    void equip(AbstractWeapon weapon);
+    void equip(IWeapon weapon);
 
     /**
      * Returns the weapon equipped by the character.
+     * @return
      */
-    AbstractWeapon getEquippedWeapon();
+    IWeapon getEquippedWeapon();
 }