+++ src/main/java/com/github/cc3002/finalreality/model/weapon/Sword.java
@@ -1,5 +1,7 @@
 package com.github.cc3002.finalreality.model.weapon;
 
+import com.github.cc3002.finalreality.model.character.player.*;
+
 public class Sword extends AbstractWeapon {
 
     /**
@@ -33,4 +35,28 @@ public class Sword extends AbstractWeapon {
     }
 
 
+    @Override
+    public void equipToKnight(Knight c) {
+        c.equippedWeapon = this;
+    }
+
+    @Override
+    public void equipToEngineer(Engineer c) {
+
+    }
+
+    @Override
+    public void equipToThief(Thief c) {
+        c.equippedWeapon = this;
+    }
+
+    @Override
+    public void equipToWhiteMage(WhiteMage c) {
+
+    }
+
+    @Override
+    public void equipToBlackMage(BlackMage c) {
+
+    }
 }