+++ src/test/java/com/github/cc3002/finalreality/model/character/AbstractPlayerCharacterTest.java
@@ -1,17 +1,16 @@
 package com.github.cc3002.finalreality.model.character;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNull;
-
 import com.github.cc3002.finalreality.model.character.player.IPlayerCharacter;
 
-import java.util.ArrayList;
-import java.util.List;
-
-import com.github.cc3002.finalreality.model.weapon.AbstractWeapon;
-import com.github.cc3002.finalreality.model.weapon.Axe;
-import org.junit.jupiter.api.Assertions;
+import com.github.cc3002.finalreality.model.character.player.Knight;
+import com.github.cc3002.finalreality.model.character.player.Thief;
+import com.github.cc3002.finalreality.model.weapon.*;
 import org.junit.jupiter.api.Test;
+
+import java.util.concurrent.LinkedBlockingQueue;
+
+import static org.junit.jupiter.api.Assertions.*;
+
 
 /**
  * Abstract class containing the common tests for all the types of player characters.
@@ -22,39 +21,29 @@ import org.junit.jupiter.api.Test;
  */
 
 public abstract class AbstractPlayerCharacterTest extends AbstractCharacterTest {
-  protected List<IPlayerCharacter> testPlayerCharacters;
-  protected AbstractWeapon testWeapon;
 
-  @Test
-  void waitTurnTest() {
-    Assertions.assertTrue(turns.isEmpty());
-    tryToEquip((IPlayerCharacter) testCharacters.get(0));
-    super.waitTurnTest();
-  }
+  public Sword sword = new Sword("sword",1, 10);
+  public Axe axe = new Axe("axe", 1,10);
+  public Knife knife = new Knife("knife", 1, 10);
+  public Staff staff = new Staff("staff", 1, 10, 1);
+  public Bow bow = new Bow("bow",1 , 10);
+
 
   /**
    * Checks that the character equips correctly.
    */
-  @Test
-  void equipWeaponTest() {
-    for (var character :
-            testPlayerCharacters) {
-      assertNull(character.getEquippedWeapon());
-      character.equip(testWeapon);
-      assertEquals(testWeapon, character.getEquippedWeapon());
-    }
+
+  protected void checkEquippedWeapon(IPlayerCharacter character, IWeapon weapon) {
+    assertNull(character.getEquippedWeapon());
+    character.equip(weapon);
+    assertEquals(weapon, character.getEquippedWeapon());
   }
 
-  private void tryToEquip(IPlayerCharacter character) {
-    character.equip(testWeapon);
+  protected void checkNotEquippedWeapon(IPlayerCharacter character, IWeapon weapon){
+    assertNull(character.getEquippedWeapon());
+    character.equip(weapon);
+    assertNull(character.getEquippedWeapon());
   }
 
-
-  protected void basicSetUp() {
-    super.basicSetUp();
-    testWeapon = new Axe("Test", 15,10);
-    testPlayerCharacters = new ArrayList<>();
-
-  }
 
 }